edge_client/requirements.txt
PyQt6
Pillow
pytesseract
mss
